requirements.txt
# Python Requirements for Robot Navigation System
# Install with: pip3 install -r requirements.txt

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0  # For ArUco markers

# Numerical Computing
numpy>=1.24.0

# Raspberry Pi Camera
picamera2>=0.3.12

# Serial Communication
pyserial>=3.5

# LIDAR Driver
rplidar-roboticia>=0.9.3

# Configuration Management
PyYAML>=6.0

# Visualization (optional)
matplotlib>=3.7.0
